# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API записной книжки",
        "description": "API для работы с контактами записной книжки. Позволяет создавать, обновлять, удалять и получать записи.",
        "contact": {
            "email": "mailMydi@yandex.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/notebook": {
            "get": {
                "tags": [
                    "Записная книжка"
                ],
                "summary": "Получить все записи",
                "description": "Возвращает список всех записей из записной книжки с постраничной навигацией",
                "operationId": "2050ac9d4f563fd2191f6a5317d34dd4",
                "responses": {
                    "200": {
                        "description": "Список записей успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Notebook"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Записная книжка"
                ],
                "summary": "Создать новую запись",
                "description": "Добавление новой записи в записную книжку",
                "operationId": "268102c45c22fe3ac8e8399768d90a2e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Иван Иванов"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+79991234567"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "ivanov@example.com"
                                    },
                                    "company": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "ООО Ромашка"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Запись успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notebook"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные входные данные"
                    }
                }
            }
        },
        "/api/v1/notebook/{id}": {
            "get": {
                "tags": [
                    "Записная книжка"
                ],
                "summary": "Получить запись по ID",
                "description": "Возвращает запись из записной книжки по её уникальному идентификатору",
                "operationId": "e0f45be39d3ac980fcd8b94e96193f9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор записи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Запись успешно найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notebook"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запись не найдена"
                    }
                }
            },
            "post": {
                "tags": [
                    "Записная книжка"
                ],
                "summary": "Обновить запись",
                "description": "Обновление существующей записи в базе данных",
                "operationId": "8d271540f8e30196109c8579b102fcf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор записи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "company": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запись успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notebook"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные входные данные"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Записная книжка"
                ],
                "summary": "Удалить запись",
                "description": "Удаление записи из базы данных",
                "operationId": "5c0d16657774a47b71087c1c92b3c707",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор записи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Запись успешно удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись успешно удалена."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запись не найдена"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Notebook": {
                "required": [
                    "name",
                    "phone",
                    "email"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "photo": {
                        "type": "string",
                        "format": "uri"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Записная книжка",
            "description": "Записная книжка"
        }
    ]
}